[#58] Review 후 수정
